Fix PlaybackErrorTrigger to use uint8_t instead of PlaybackError enum
The automation system expects Trigger<uint8_t> (matching the Python cg.uint8),
not Trigger<PlaybackError>. Changed the trigger class, callback signature, and
error call site to use uint8_t with proper casting.

+++ esphome/components/simple_video_player/simple_video_player.cpp
@@ -685,7 +685,7 @@ void SimpleVideoPlayer::set_error_(PlaybackError error) {
   this->state_ = PlayerState::ERROR;
   xSemaphoreGive(this->state_mutex_);
 
-  this->on_error_callbacks_.call(error);
+  this->on_error_callbacks_.call(static_cast<uint8_t>(error));
 }
 
 }  // namespace esphome::simple_video_player
